[iOS] refactor :recycle: IssueLabelsCollectionView ViewController로 변경
- collection view가 많은 일을 하지 않도록 변경
- IssueHorizontalCell마다 하나의 view controller 갖도록 구현

#70
